pyproject.toml
[tool.poetry]
name = "tableau"
version = "0.1.0"
description = ""
authors = ["duong <duong.nguyen2@decathlon.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.14"
tableauserverclient = "^0.38"
pandas = "^2.3.3"
tableauhyperapi = "^0.0.24081"
pyarrow = "^23.0.0"
fastparquet = "^2025.12.0"
databricks-sql-connector = "^4.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
